Extract the `Input` component to a separate file.
